+++ src/LayoutUtils.ts
@@ -17,7 +17,7 @@ class LayoutUtils {
             lines.pop();
 
         // calculate minimum indent which can be found at every line
-        var indentLens = lines.map(line => /^ */.exec(line)[0].length);
+        var indentLens = lines.map(line => /^ */.exec(line)[0].length).filter(x => x !== 0);
         var indentLen = Math.min(...indentLens);
         if (indentLen)
             lines = lines.map(line => line.substr(indentLen));
@@ -25,13 +25,14 @@ class LayoutUtils {
         return lines.join('\n').trim();
     }
 
-    static addEditor(name, lang?: string, onChanged?: (origin:string, newValue:string) => void) {
+    static addEditor(name, lang ?: string, onChanged?: (origin: string, newValue: string) => void, isReadOnly: boolean = false) {
         var editorName = `${name}Editor`;
         var editor = qxSchema.ui[editorName] = ace.edit(editorName);
         editor.setTheme("ace/theme/monokai");
         if(lang)
             editor.getSession().setMode(`ace/mode/${lang}`);
         editor.setOption('tabSize', 2);
+        editor.setReadOnly(isReadOnly);
         editor.$blockScrolling = Infinity; // TODO: remove this line after they fix ACE not to throw warning to the console
 
         var localStorageKey = `model_${name}_value`;
@@ -42,7 +43,9 @@ class LayoutUtils {
         var editorDelay = new Delayed(750);
         editor.on('change', () => editorDelay.do(() => {
             var newValue = editor.getValue();
-            localStorage.setItem(localStorageKey, newValue);
+            if (!isReadOnly)
+                localStorage.setItem(localStorageKey, newValue);
+
             qxSchema.model.editors[name] = newValue;
             onChanged && onChanged(name, newValue);
         }));
